all changes

+++ app/src/main/java/com/example/fyp/Account.kt
@@ -3,7 +3,7 @@ package com.example.fyp
 data class Account(
     var id: String = "",
     var accName : String = "",
-    var accCardNumber : Int = 0,
+    var accCardNumber : Long = 0,
     var accCardDate : Int = 0,
     var accCardCode : Int = 0,
     var accCardAmount : Double = 0.0,
